reorganize shaders; use modulate flag; update librw
