+++ src/extras/screendroplets.cpp
@@ -112,14 +112,14 @@ ScreenDroplets::InitDraw(void)
 
 	openim2d_uv2();
 #ifdef RW_D3D9
-#include "shaders/screenDroplet_PS.inc"
+#include "shaders/obj/screenDroplet_PS.inc"
 	screenDroplet_PS = rw::d3d::createPixelShader(screenDroplet_PS_cso);
 #endif
 #ifdef RW_GL3
 	using namespace rw::gl3;
 	{
-#include "shaders/im2d_UV2_gl.inc"
-#include "shaders/screenDroplet_fs_gl.inc"
+#include "shaders/obj/im2d_UV2_gl.inc"
+#include "shaders/obj/screenDroplet_fs_gl.inc"
 	const char *vs[] = { shaderDecl, header_vert_src, im2d_UV2_vert_src, nil };
 	const char *fs[] = { shaderDecl, header_frag_src, screenDroplet_frag_src, nil };
 	screenDroplet = Shader::create(vs, fs);
