+++ src/extras/shaders/makeinc_hlsl.sh
@@ -1,4 +1,5 @@
 #!sh
+cd obj
 for i in *cso; do
 	(echo -n 'static '
 	xxd -i $i | grep -v '_len = ') > ${i%cso}inc
