--- src/extras/shaders/simple_fs_gl.inc
@@ -1,19 +1,0 @@
-const char *simple_frag_src =
-"uniform sampler2D tex0;\n"
-
-"FSIN vec4 v_color;\n"
-"FSIN vec2 v_tex0;\n"
-"FSIN float v_fog;\n"
-
-"void\n"
-"main(void)\n"
-"{\n"
-"	vec4 color;\n"
-"	color = v_color*texture(tex0, vec2(v_tex0.x, 1.0-v_tex0.y));\n"
-"	color.rgb = mix(u_fogColor.rgb, color.rgb, v_fog);\n"
-"	DoAlphaTest(color.a);\n"
-
-"	FRAGCOLOR(color);\n"
-"}\n"
-
-;
